Game.java
Main class for the Connect 4 game. Utilizes two other classes I made "Slot" and "Player" A two-dimensional array of "Slots" is created and each "Slot" can hold a symbol like an 'x' or 'o'. The Player class holds a bunch of information such as how many chips remain and includes a method for dropping chips into the board.